Inter-nucleus distance calculations
For the diffusion equation, I need inter-nuclear distances.  Started on
the code for calculating that.  It may even work.  Had to reduce
position to 32 bit floats to get it to calculate properly (otherwise, we
run out of memory).  The piece-wise calculation of the distance matrix
is for a related Array Too Big error.