Fix bug on comparison
Since has_difference was basically a string comparison, I've deleted it
and replaced it with a string comparison.  Also, I realized I wasn't
properly handling gaps, so 'AA-T' and 'AA-T' would get marked as
different, due to the gaps getting replaced.

The only time when gaps are actually "different" from each other are
when the whole sequence is gaps, in which case it should be ignored.
Since 'Both sequences are all gaps' is a subset of "both sequences are
the same", I've got the sort circuit and to break out when the sequences
are different. I should look and see whether there's a more efficient
way of doing this particular line, such that it breaks out as soon as
there's a non-gap.