Change dictionary to index on sequence and fix algorithm
Indexing on sequence should allow us to use the filtered_hits.sam file
instead of accepted_hits, which is (unfortunately), an order of
magnitude smaller.

Also, the way I look for lines to keep should actually work now, as
opposed to before, when it would see a non-ID line and deicide that it
didn't match.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# On branch PAC03
# Changes to be committed:
#   (use "git reset HEAD <file>..." to unstage)
#
#	modified:   FindUnmappedReads.py
#