Switch back to using read ID as key
For reasons that still don't make complete sense to me, using the
sequence of even accepted_hits.sam finds some 22 million unmapped reads,
which is about twice as large as it should be.  Switching back to the
read name does better (11 million reads or so), which is still larger
than it should be, but at least in the right ballpark.
# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# On branch PAC03
# Changes to be committed:
#   (use "git reset HEAD <file>..." to unstage)
#
#	modified:   FindUnmappedReads.py
#