Separate out bowtie and express
Now has all the post-processing of the bowtie samfile so that I dont
have to pipe it directly into express, and critically so it can do the
second round of EM.

Also, I've removed the actual call to express so this can run on deme.
I'll put it back in at some point.