Regularize data in FindDifferenceCounts
Data as is may not be the same shape, and therefore can't be made into
an array.  Fortunately, the structure of the code in TSSFinder is such
that if a column is missing, it's always going to be at the end, so I
can just pad with zeros.

Except I can't just pad with zeros, so I introduce pseudo-counts.  Who
knows if that's actually the correct thing to do, and will introduce
errors in the analysis downstream in the case of zero-expression in the
middle.  Meh.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# On branch PAC06
# Changes to be committed:
#   (use "git reset HEAD <file>..." to unstage)
#
#	modified:   FindDifferenceCounts.py
#
# Untracked files:
#   (use "git add <file>..." to include in what will be committed)
#
#	TSSFinder.py