Radically speedup TSSFinder.py
Set up a dictionary with all chrom, basepair pairs in a first exon, and use
that to determine whether a read falls in the first exon, rather than searching
through first exons.  This requires a somewhat larger memory footprint (not
sure how large it gets, but it seems to be manageable on deme), but goes much,
much faster.