Use NaN for points with no data
A thought that's been niggling at me for a while is that MatchSlices
doesn't do a great job of getting back the position of the slice, even
in relatively good situations.  One possibility I've been toying with is
the fact that for lots of genes, there isn't data at all the time
points.  Maybe that is one thing that's making it hard to place the
slices in time (if not in space).  Therefore, I've modified the code
such that all the data is NaNs, except for data that is explicitly put
there (even if that data does turn out to be zero), whereas before I was
assuming everything was zero.  Then, when doing the spearman
correlation, I ignore any entry that has a NaN. This does, in fact, seem
to help with placing things in time (at least it gives more reasonable
numbers for all the time points), but makes very little difference in
space.