Add "MatchToSusan.py"
A (still hacky) script to figure out which of Susan Lott's RNA-seq time
points my RNA-seq data looks most like.  Does this by calculating a
log-likelihood score, whereby the expression is assumed to be normally
distributed around the mean of the all embryos at that time point, with
std as the std of all those expression values.

Yes, I'm aware that this is anywhere from 2-4 embryos per time point, so
not great.  Yes, I'm aware that my embryo might not be from any of those
time points. Yes, I'm aware that making the assumption that it's
normally distributed is likely to be pretty crappy anyways. But it's a
start.