Changes for PAC40 specific run
Add in the use of AssignReads2, which will attempt to figure out which species
each read came from, or give an indication if it can't be assigned
unambiguously.  Upon reflection, there may be a problem with the output order
getting mingled in with everything else... maybe I should pipe and save it for
later?

Selection of indices, choice of right reference files.