Create AssignReads2
This works by using the NM and NH tags in the samfile to make assignment calls.
The first version of this that I wrote goes by looking at the 0x100 flag, which
according to the documentation, corresponds to a read being the secondary
alignment. I'm blown away that ever worked, and if it did, it seems like it'd
be super conservative.   This works by keeping track of the best-mapping
location for each read from each species (as defined by edit distance), then
when all the mappings for that read are in, it figures out which ones can be
unambiguously assigned.  It also attempts to figure out how large the
delta-edit-distance between the first and second best species is.