Hunt Dark Matter now incorporates gene expression levels
Previously, the null model was that all the slices would be equal, which
is stupid.  Now, we're assuming that all the slices use each enhancer in
proportion to the total expression, which is clearly wrong, but not a
stupid null model.

Also add in BigWig handling (should work interchangeably with .bam
files, but has never been tested), and an ability to use SIGTERM to skip
the current gene (but not kill the whole program).  SIGKILL still brings
us to a grinding halt.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# On branch PAC45
# Changes to be committed:
#   (use "git reset HEAD <file>..." to unstage)
#
#	modified:   HuntDarkmatter.py
#
# Untracked files:
#   (use "git add <file>..." to include in what will be committed)
#
#	PlotSigDarkmatter.py