Remove the plotting from do_tux
For one thing, it doesn't work on deme, for another, I don't actually care
about it any more, and for a third, it no longer (if ever) really logically fit
in with the tuxedo suite, but instead is its own idea.  Will eventually make a
separate script that makes some/all of the figures.