Spin off the read mapping to a separate script
In the big data run, I'm going to want to have a separate configuration
file, as well as not run the cuffdiff until I can assign some sort of
reasonable "replicates", depending on exactly what the hell I mean by
that. So I separate the cuffdiff call until later, and in the mean time,
simplify all the read maping such that it should work sanely.