Move config variables into a holder object
This is done to inflate my pylint scores.  Otherwise, should have no
effect on the shape of the code.

The one change I've made to the function is to calculate the sequence
directory, then add in the 'R1' or 'R2', to be explicit that both ends
should be in the same directory.