Only do cufflinks on Base species, and other pylint fixes
Use the base_GTF argument to give the path to the GTF for the base
species only.  I suspect this will speed up the cufflinksing, which
seems to be a moderate bottleneck, at least in the dry-run testing I've
done.

I've also moved more variables into their own Namespaces, creating a
TEMP namespace for names that get reusused willy-nilly.