Find readfiles on the fly
Rather than getting all the readfiles in advance, for each sample in the
config file, calculate the read files for that sample at the last
possible moment (i.e. before doing tophat).  This will allow us to start
doing the tophat as the files are being downloaded, which could save a
few hours.