Various scripts to fix the assigned reads
RescueAmbiguous looks for read pairs where one end is unambiguously
assigned, and where the other is not, and uses the unambiguous one to
"rescue" the ambiguous one.

FixAssignedBam goes through the reads and fixes the various SAM columns
based on which ones actually got unambiguously assigned, rather than
what the mapper said.

CalculateConcentrations reads a tsv file with the read numbers and
carrier concentrations, and estimates the original D. mel concentration.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# On branch Fall2012
# Your branch is ahead of 'origin/Fall2012' by 2 commits.
#
# Changes to be committed:
#   (use "git reset HEAD <file>..." to unstage)
#
#	new file:   CalculateConcentrations.py
#	new file:   FixAssignedBam.py
#	new file:   RescueAmbiguous.py
#	modified:   Utils.py
#