Use Bayes theorem to assign each sample to a point in time and space
This is part of a storied attempt to localize each of the samples to a point in
space (using the BDTNP dataset) and a point in time (using Lott 2011's single
embryo set).  I don't know why it took me this long to invoke Bayes' theorem,
which is all about managing conflicting sets of probabilities.