In CalculateAbundances, do a No-bcd run
Due to multiple testing, and especially due to the fact that it goes up
by the square of the number of things we're testing, including the bcd
samples in one round really kills our ability to detect.

Furthermore, it looks like we're not going to look at the bicoid data in
the paper.