Track paired-end reads properly, and allow multiple sequence directories
At some point, I changed the search string for the first end, but not
the second, and that meant that I was only doing single-end mapping. As
part of this, I include a warning (technically, just a print
statement... maybe this should go to stderr) if it doesn't find the
second end as well. Also, I sort both ends, so there's less room for
error if things get globbed in different orders (this appeared to have
been a problem)

Also, since the first round of sequencing the data in this batch didn't
work, I have a sequence and a sequence2 directory. Adding a * in the
glob path after sequence now catches both.