Allow configure to pool adjacent slices
By making a two-stage process, first reading in each line of the
configuration file, then only writing out to the makefile when I know
all the reads in that label.

Note that this does not preserve the yeast identities, so it is not
suitable for de-pooling the yeast from the combined samples, but this
also shouldn't delete directories, so I can presumably switch between
them without much loss of data.