Fix how I deal with yeast
SGD labels the fasta file one way, and the GFF file a different way, and
so ReheaderYeast should (I hope) help reconcile that.

Also, the changes to allow multiple samples get pooled in analysis broke
the way I was handling the yeast in the config file. That *should* now
be fixed.