Makefile and configure script now honor genome
In the past, I just mapped everything to yeast, but this is really much
better.  I think I caught everywhere it matters.

I had to stop using samtools reheader to fix the header, since I
couldn't know that the chromosomes would be in the same order (and, for
the flies, they aren't)