Fixed a bug with no index, return values
If the bam file is not indexed, it won't be able to count the reads
properly, so it just goes through all of them once.  However, seek()ing
back to the beginning didn't work, so I'm now just re-opening the Bam
file.

I've also included return values from the sub-functions, so that I can
more closely scrutinize the results.