Code to make a table of ISH images
I made this more than a year ago, and have tweaked it since. It *should*
be able to work in both Python2 and Python3.  It uses a GUI to select
different images, which is probably why I hadn't committed it until now,
since most of the code in my various git repositories are for command
line use.