MakeSummaryTable gets read files from bam header
This is way better than trying to guess it based on the constantly
changing formats the sequencing center gives the data in.

There's a bit of a risk still present in that it doesn't count all the
reads, so if there are multiple "leftovers" files (for instance, if
adjacent slices are computationally merged), then it won't deal with
them properly.