No plan survives contact with data
As expected, I had to add in a new glob pattern for the data as it came
back.

Also dropping the mapping rate requirement.  As the data came in, there
were some empty files from the other lanes that threw off the counting
of unmapped reads.  May put this back in once re-mapped with the proper
glob specification.