Refactor how I calculate the HSV of the multicolor ISH CMs
Now I'm making it a bit more explicit that the Hues are spread evenly
throughout the color wheel (0-1, N even steps), whereas the Saturation
and Value are the same between all.

This allowed me to more easily make CM sets with different numbers of
colors, all containing the in situ blue.