Improve code for making averages
Major step is the "draw_average_only" option, which as the name
suggests, does not draw the complete heat map, only the averages.

Also a feature for scaling the averages.  This seems to be necessary
because not all patterns are going to be the same, so there's some
amount of dilution of the patterns. This approach increases the contrast
by a user-definable factor (but default of 1)