Split jobs among processors by splitting, not blocking
i.e.
CPU1: 1  4  7  10
CPU2:  2  5  8
CPU3:   3  6  9

Not
CPU1: 123
CPU2:    456
CPU3:       78910

My assumption here is that jobs that are close to each other in sequence
are likely to be more similar---in length, likelihood to change
independently, etc---so having them split among different processors is
more likely to be equitable.

Also, I think the code is a bit cleaner, and guarantees that everything
gets run, which the other you could imagine that things at the end get
left out.