Script for packaging files for GEO
I had the realization that when making the summary table, I have
knowledge both of where the bam file is (which in turn has a link to the
reads that go into it) and what the name of the column is. Thus, I can
just write out a Makefile that contains both of those in a linked way,
and a script that reads in the bam file and writes out the correct name.
It'll take a while to run, but it's not so bad.